Create sequence to generate wrapped cor and app problem with marxan on both
This only runs up through running marxan on both problems, but it runs to completion
and looks like it's working that far.

One thing is that it does NOT run marxan on the base cor problem.
I would have to extract the base cor problem pointer from the wrapped problem and
run marxan on that.  I've dummied in a bit of code describing what would have to
go there, but not the actual function needed and not implemented yet.
This would probably be much easier to do if there was a library of generated objects
with an index somewhere and objects were added to that library and index as soon
as they were generated.  Even then, there's another question about how well do the
directory structures hold up if getting things from a library?  For example,
when you create an app structure, it relies on the cor structure, but that might
have been generated and existed in a previous run, but not this run.  The dir
tree builders need to be robust to that.

Another question about that, would it correctly decide where to put the marxan
results?