Tidying, but entire analysis runs to competion now for wrap and apparent
The entire analysis runs to completion for wrapping and an apparent problem.
However, it needs a lot of careful examination of the output to see whether
it's doing things correctly.
It also needs lots of refactoring of the bdpg code for reading marxan files
and for writing the final output, since both of those are a big, long mess
of code.