Refactor main body of code into function to prepare for different main options
About to start building different ways of loading and building problem
objects to allow for something more like a pipeline.  The main code
that's been here always builds a base COR, base APP, wrap COR, and wrap APP
unless some of them are turned off.  Moving this set of operations into a
single function that will probably not be used in the long run, but still
works now.