Tidying.  Not sure if this function will still work after all the recent changes.
Have just finished replacing much of this function with single action function
call.  Will probably not use this function anymore, but it might still be
handy for occasional one-off tests, so I've tidied it up in the last few
commits so that it might still work.  Need to test it.