Temporarily stop turning warnings into errors.
Having a problem on nectar under ubuntu 16.04 with a warning turning into an
error and crashing everything:

Error in FUN(X[[i]], ...) :
  (converted from warning) failed to assign RegisteredNativeSymbol for toeplitz to toeplitz since toeplitz is already defined in the ‘spam’ namespace

This warning does not happen on mac and is
probably due to some problems with having to use old versions of R packages to
get something working under this old version of ubuntu.
Once we get tzar and ubuntu etc all working correctly on nectar, then this
can be changed back to turning warnings into errors, I think.