Start new project & emulation for testing loading existing tzar runs from other machines.
This is the beginning of building and testing functions in bdpg file
load_existing_tzar_run.R.  The goal is to be able to scp old tzar runs from another
machine and then act on those existing runs (e.g., to run a different reserve selector
on an existing problem or run graph calculations on an existing problem).
This is useful in allowing nectar runs on machines of different sizes in different
passes since some actions use little memory and can use lots of machines while others
need lots of memory, etc.